✨feature: nouvel gestion des erreur

+++ backend/app/utils/generatController.js
@@ -1,4 +1,4 @@
-import ApiError from '../error/apiError.js';
+import { ApiError } from '../error/apiError.js';
 /**
  *
  * @param {*} model Model Sequelize
@@ -12,9 +12,13 @@ const generateController = (model) => ({
   async getOne(req, res) {
     const id = Number(req.params.id);
 
+    if (!id) {
+      throw new ApiError('Invalid parameter id', `${model.name}Error`);
+    }
+
     const data = await model.findByPk(id);
     if (!data) {
-      throw new ApiError(400, `No ${model.name} with id ${id} found`);
+      throw new ApiError(`This ${model.name} does not exist`, `${model.name}Error`, 0);
     }
 
     res.json(data);
@@ -25,30 +29,42 @@ const generateController = (model) => ({
     res.status(201).json(data);
   },
   async update(req, res) {
+    const id = Number(req.params.id);
+
+    if (!id) {
+      throw new ApiError('Invalid parameter id', `${model.name}Error`);
+    }
+
     const [nbUpdated, dataUpdated] = await model.update(req.body, {
 
       where: {
-        id: req.params.id,
+        id,
       },
 
       returning: true,
     });
 
     if (nbUpdated === 0) {
-      throw new ApiError(400, `No ${model.name} with id ${req.params.id} found`);
+      throw new ApiError(`This ${model.name} does not exist`, `${model.name}Error`, 0);
     }
 
     res.json(dataUpdated[0]);
   },
   async delete(req, res) {
+    const id = Number(req.params.id);
+
+    if (!id) {
+      throw new ApiError('Invalid parameter id', `${model.name}Error`);
+    }
+
     const nbDeleted = await model.destroy({
       where: {
-        id: req.params.id,
+        id,
       },
     });
 
     if (!nbDeleted) {
-      throw new ApiError(400, `No ${model.name} with id ${req.params.id} found`);
+      throw new ApiError(`This ${model.name} does not exist`, `${model.name}Error`, 0);
     }
 
     // 204 => On retourne rien mais tout c'est bien passé
